refactor: extract scout-runner and swarm-utils from swarm.ts
